+++ src/database/queries/changemakers/insertOne.sql
@@ -1,0 +1,8 @@
+INSERT INTO changemakers (
+  tax_id,
+  name
+) VALUES (
+  :taxId,
+  :name
+)
+RETURNING changemaker_to_json(changemakers) AS "object";