+++ src/database/queries/changemakers/selectByTaxId.sql
@@ -1,0 +1,3 @@
+SELECT changemaker_to_json(changemakers) AS "object"
+FROM changemakers
+WHERE tax_id = :taxId