+++ src/database/queries/changemakers/selectWithPagination.sql
@@ -1,7 +1,7 @@
 SELECT DISTINCT o.id,
-  organization_to_json(o.*, :keycloakUserId) AS "object"
-FROM organizations o
-  LEFT JOIN organizations_proposals op on op.organization_id = o.id
+  changemaker_to_json(o.*, :keycloakUserId) AS "object"
+FROM changemakers o
+  LEFT JOIN changemakers_proposals op on op.changemaker_id = o.id
 WHERE
   CASE
     WHEN :proposalId::integer IS NULL THEN