+++ src/database/queries/proposals/selectWithPagination.sql
@@ -2,7 +2,7 @@ SELECT proposal_to_json(p.*) as "object"
 FROM proposals p
   LEFT JOIN proposal_versions pv ON pv.proposal_id = p.id
   LEFT JOIN proposal_field_values pfv on pfv.proposal_version_id = pv.id
-  LEFT JOIN organizations_proposals op on op.proposal_id = p.id
+  LEFT JOIN changemakers_proposals op on op.proposal_id = p.id
 WHERE
   CASE
     WHEN :createdBy::UUID IS NULL THEN
@@ -18,10 +18,10 @@ WHERE
       pfv.value_search @@ websearch_to_tsquery('english', :search::text)
     END
   AND CASE
-    WHEN :organizationId::integer IS NULL THEN
+    WHEN :changemakerId::integer IS NULL THEN
       true
     ELSE
-      op.organization_id = :organizationId
+      op.changemaker_id = :changemakerId
     END
   AND CASE
     WHEN :authContextKeycloakUserId::UUID IS NULL THEN