+++ src/database/queries/sources/insertOne.sql
@@ -2,12 +2,12 @@ INSERT INTO sources (
   label,
   data_provider_short_code,
   funder_short_code,
-  organization_id
+  changemaker_id
 )
 VALUES (
   :label,
   :dataProviderShortCode,
   :funderShortCode,
-  :organizationId
+  :changemakerId
 )
 RETURNING source_to_json(sources) AS "object";